requirements.txt
anthropic>=0.40.0
requests>=2.31.0
apscheduler>=3.10.0
python-dotenv>=1.0.0
openai>=1.0.0
yfinance>=1.2.0
